gamemodreadme
Rainbow Six Siege Game Mod ReadMe

The mod for the midterm is located in midtermBranch
To play the mod the best thing to do is upon starting a new game, once the player is on the screen hit Ctrl, Alt, ~ to open up the console and type in one of the following commands:
    - give class_1
    - give class_2
    - give class_3
    - give class_4
These are the custom classes of the mod.  This is also how to show the deliverable of 4 classes as well as Quake 4 weapons that behave as Rainbow6Siege weapons.  After that you can play the game as normal and you will notice that it might be a lot more difficult to survive compared to normal Quake 4 which was the point of this mod.  Which was what i wanted to aim for with this mod.  The following is a way to test the other deliverables of the mod.

Item drops and 5 unique weapon drops. To see this deliverable, kill an enemy and then it will drop an item.  I set all the new drops to be dropped from all enemies but to        see them better, open the console and type in:
    - spawn item_health_hurt50  (this does 50 damage to the player, looks like a small health pack)
    - spawn item_health_death  (this instantly kills the player, looks like a small healthpack)
    -spawn item_health_nothing  (this is a healthpack that does no healing at all to the player)
    - spawn empty_mg_clip  (this is a machine gun pickup that doesn't give any bullets at all)
    - spawn goodbye_mg_clip  (this is a machine gun pickup that takes away 15 bullets from the reserve from your ammo)
*NOTE for the last 2 machine gun items, use give class_1 in the console to see the changes made with the items on the machine gun

Damage Multiplier deliverable.  To see this deliverable you can keep the blaster you don't need to go into a class to see it.  The blaster I have set to 100 damage, but the strogg marine which is the first enemy in the game will take no headshot damage.  If you shoot it's head it will not die no matter what.  To test go into the console and type in the following in this order:
    -notarget (so the enemies won't target you)
    -spawn monster_strogg_marine
    - and just wait for it to stop moving and fire the NPCS, you will have an opportunity to easily shoot the enemy in the head and it wont die.  If you shoot it in the chest it will die.

Deliverable I did not complete was the adding of more enemies in the level.